feat(BankOfSimba) part2

+++ week_07/day_3/Bank of Simba/Bank of Simba/Models/BankAccount.cs
@@ -3,7 +3,7 @@
 	public class BankAccount
 	{
 		public string Name { get; set; } = "Simba";
-		public int Balance { get; set; } = 2000;
+		public double Balance { get; set; } = 2000;
 		public string AnimalType { get; set; } = "Lion";
 	}
 }
